pin movement check for Unklicky TAP instructions
The pin needs to be able to move freely in the body part for safe operation.
This may not be the case if the printed parts have slight deficiencies.
These can include overextruded top or coarse bottom layers (textured PEI).

This commit adds a relevant safety check to the build instructions.
It also provides a tip about sanding pin parts.

Signed-off-by: Tilmann Unte <tilmann.unte@googlemail.com>